dirt: compile-time precision abstraction (Real/Accum) + soil@main
Thread soil's precision-double/mixed/single feature through every dirt crate
(mirroring soil's Real/Accum design) and cast at the f32/f64 boundaries in the
contact, wall, bond, clump, fixes and measure-plane kernels. Repin all soil
dependencies from branch="master"/"dev/gpu" to branch="main".

+++ crates/dirt_granular/src/rotational.rs
@@ -130,7 +130,7 @@ mod tests {
         atom.dt = dt;
 
         push_dem_test_atom(&mut atom, &mut dem, 0, [0.0; 3], radius);
-        let mass = atom.mass[0];
+        let mass = atom.mass[0] as f64;
         let inertia = 0.4 * mass * radius * radius;
 
         // Apply torque around z-axis
